all code without slurm/spank symbols moved to lib path
