+++ src/lib/database.cpp
@@ -4,9 +4,8 @@
 #include <vector>
 
 #include "database.hpp"
+#include "filesystem.hpp"
 #include "sqlite.hpp"
-
-#include "../util/filesystem.hpp"
 
 namespace db {
