+++ src/lib/parse_args.hpp
@@ -4,23 +4,18 @@
 #include <string>
 #include <vector>
 
-#include "util/expected.hpp"
+#include <lib/expected.hpp>
+#include <lib/mount.hpp>
 
-namespace impl {
+namespace util {
 
 enum class protocol { file, https, jfrog };
 
-struct mount_entry {
-  // enum protocol p;
-  std::string image_path;
-  std::string mount_point;
-};
-
 /// parse list of `:` separated tuples, enforce absolute paths, sort output by
 /// mountpoint. The error state is an error string.
-util::expected<std::vector<mount_entry>, std::string>
+util::expected<std::vector<util::mount_entry>, std::string>
 parse_arg(const std::string &arg,
           std::optional<std::string> uenv_repo_path = std::nullopt,
           std::optional<std::string> uenv_arch = std::nullopt);
 
-} // namespace impl
+} // namespace util
