+++ src/lib/strings.cpp
