+++ src/lib/strings.hpp
